+++ mobile/android/fenix/app/src/test/java/org/mozilla/fenix/share/ShareControllerTest.kt
@@ -597,7 +597,7 @@ class ShareControllerTest {
 
             every {
                 sendTabUseCases.sendToDeviceAsync(
-                    any(),
+                    any<String>(),
                     any<List<TabData>>(),
                 )
             } returns CompletableDeferred(true)