+++ mobile/android/android-components/docs/changelog.md
@@ -5,6 +5,8 @@ permalink: /changelog/
 ---
 
 # 158.0 (In Development)
+* **feature-accounts-push**
+    * 🆕 Added a `SendTabUseCases.SendToDeviceUseCase` overload that takes a list of devices, sending every tab to every device. [Bug 2056922](https://bugzilla.mozilla.org/show_bug.cgi?id=2056922)
 * **service-pocket**
     * ⚠️ **Breaking change**: Removed `useMerinoClient` from `ContentRecommendationsRequestConfig`. The content recommendations are always fetched with the Merino client. [Bug 2069992](https://bugzilla.mozilla.org/show_bug.cgi?id=2069992)